refactored, renaming application. Added privacypolicy file.
