+++ app/src/androidTest/java/com/jrmz/taskmaster/AddNewTaskVisibleOnHomeActivity.java
@@ -1,8 +1,10 @@
-package com.example.taskmaster;
+package com.jrmz.taskmaster;
 
 
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.action.ViewActions.click;
+import static androidx.test.espresso.action.ViewActions.closeSoftKeyboard;
+import static androidx.test.espresso.action.ViewActions.replaceText;
 import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
@@ -15,12 +17,16 @@ import android.view.ViewGroup;
 import android.view.ViewParent;
 
 import androidx.test.espresso.ViewInteraction;
+import androidx.test.espresso.matcher.ViewMatchers;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.LargeTest;
 
+import com.example.taskmaster.R;
+
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
+import org.hamcrest.Matchers;
 import org.hamcrest.TypeSafeMatcher;
 import org.junit.Rule;
 import org.junit.Test;
@@ -28,47 +34,59 @@ import org.junit.runner.RunWith;
 
 @LargeTest
 @RunWith(AndroidJUnit4.class)
-public class HomeActivityTestAllTasksDisplays {
+public class AddNewTaskVisibleOnHomeActivity {
 
     @Rule
     public ActivityScenarioRule<HomeActivity> mActivityScenarioRule =
             new ActivityScenarioRule<>(HomeActivity.class);
 
     @Test
-    public void homeActivityTestAllTasksDisplays() {
-        // Added a sleep statement to match the app's execution delay.
-        // The recommended way to handle such scenarios is to use Espresso idling resources:
-        // https://google.github.io/android-testing-support-library/docs/espresso/idling-resource/index.html
-        try {
-            Thread.sleep(500);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-
+    public void addNewTaskVisibleOnHomeActivity() {
         ViewInteraction materialButton = onView(
-                allOf(withId(R.id.homeAllTasksButton), withText("All Tasks"),
+                Matchers.allOf(ViewMatchers.withId(R.id.homeAddTaskButton), withText("Add Task"),
+                        childAtPosition(
+                                childAtPosition(
+                                        withId(android.R.id.content),
+                                        0),
+                                1),
+                        isDisplayed()));
+        materialButton.perform(click());
+
+        ViewInteraction appCompatEditText = onView(
+                allOf(withId(R.id.addTaskTaskTitleText),
                         childAtPosition(
                                 childAtPosition(
                                         withId(android.R.id.content),
                                         0),
                                 2),
                         isDisplayed()));
-        materialButton.perform(click());
+        appCompatEditText.perform(replaceText("Feed Dawg"), closeSoftKeyboard());
 
-        // Added a sleep statement to match the app's execution delay.
-        // The recommended way to handle such scenarios is to use Espresso idling resources:
-        // https://google.github.io/android-testing-support-library/docs/espresso/idling-resource/index.html
-        try {
-            Thread.sleep(700);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
+        ViewInteraction appCompatEditText2 = onView(
+                allOf(withId(R.id.addTaskTaskDescriptionText),
+                        childAtPosition(
+                                childAtPosition(
+                                        withId(android.R.id.content),
+                                        0),
+                                4),
+                        isDisplayed()));
+        appCompatEditText2.perform(replaceText("He is starving!"), closeSoftKeyboard());
+
+        ViewInteraction materialButton2 = onView(
+                allOf(withId(R.id.addThisTaskButton), withText("Submit"),
+                        childAtPosition(
+                                childAtPosition(
+                                        withId(android.R.id.content),
+                                        0),
+                                5),
+                        isDisplayed()));
+        materialButton2.perform(click());
 
         ViewInteraction textView = onView(
-                allOf(withId(R.id.textView7), withText("All Tasks"),
-                        withParent(withParent(withId(android.R.id.content))),
+                allOf(withId(R.id.fragTaskListFragment), withText("1:\n Feed Dawg [NEW]: \nHe is starving!"),
+                        withParent(withParent(withId(R.id.homeTaskListRecyclerView))),
                         isDisplayed()));
-        textView.check(matches(withText("All Tasks")));
+        textView.check(matches(withText("1:  Feed Dawg [NEW]:  He is starving!")));
     }
 
     private static Matcher<View> childAtPosition(
