+++ app/src/androidTest/java/com/jrmz/taskmaster/SetUserNameRetainedOnHomeActivity.java
@@ -1,7 +1,8 @@
-package com.example.taskmaster;
+package com.jrmz.taskmaster;
 
 
 import static androidx.test.espresso.Espresso.onView;
+import static androidx.test.espresso.Espresso.pressBack;
 import static androidx.test.espresso.action.ViewActions.click;
 import static androidx.test.espresso.action.ViewActions.closeSoftKeyboard;
 import static androidx.test.espresso.action.ViewActions.replaceText;
@@ -17,12 +18,16 @@ import android.view.ViewGroup;
 import android.view.ViewParent;
 
 import androidx.test.espresso.ViewInteraction;
+import androidx.test.espresso.matcher.ViewMatchers;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.LargeTest;
 
+import com.example.taskmaster.R;
+
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
+import org.hamcrest.Matchers;
 import org.hamcrest.TypeSafeMatcher;
 import org.junit.Rule;
 import org.junit.Test;
@@ -30,59 +35,51 @@ import org.junit.runner.RunWith;
 
 @LargeTest
 @RunWith(AndroidJUnit4.class)
-public class AddNewTaskVisibleOnHomeActivity {
+public class SetUserNameRetainedOnHomeActivity {
 
     @Rule
     public ActivityScenarioRule<HomeActivity> mActivityScenarioRule =
             new ActivityScenarioRule<>(HomeActivity.class);
 
     @Test
-    public void addNewTaskVisibleOnHomeActivity() {
+    public void setUserNameRetainedOnHomeActivity() {
         ViewInteraction materialButton = onView(
-                allOf(withId(R.id.homeAddTaskButton), withText("Add Task"),
+                Matchers.allOf(ViewMatchers.withId(R.id.homeUserSettingsButton), withText("User Settings"),
+                        childAtPosition(
+                                childAtPosition(
+                                        withId(android.R.id.content),
+                                        0),
+                                3),
+                        isDisplayed()));
+        materialButton.perform(click());
+
+        ViewInteraction appCompatEditText = onView(
+                allOf(withId(R.id.settingsEditTextPersonName),
                         childAtPosition(
                                 childAtPosition(
                                         withId(android.R.id.content),
                                         0),
                                 1),
                         isDisplayed()));
-        materialButton.perform(click());
+        appCompatEditText.perform(replaceText("Jon"), closeSoftKeyboard());
 
-        ViewInteraction appCompatEditText = onView(
-                allOf(withId(R.id.addTaskTaskTitleText),
+        ViewInteraction materialButton2 = onView(
+                allOf(withId(R.id.settingsSubmitButton), withText("Submit"),
                         childAtPosition(
                                 childAtPosition(
                                         withId(android.R.id.content),
                                         0),
                                 2),
                         isDisplayed()));
-        appCompatEditText.perform(replaceText("Feed Dawg"), closeSoftKeyboard());
-
-        ViewInteraction appCompatEditText2 = onView(
-                allOf(withId(R.id.addTaskTaskDescriptionText),
-                        childAtPosition(
-                                childAtPosition(
-                                        withId(android.R.id.content),
-                                        0),
-                                4),
-                        isDisplayed()));
-        appCompatEditText2.perform(replaceText("He is starving!"), closeSoftKeyboard());
-
-        ViewInteraction materialButton2 = onView(
-                allOf(withId(R.id.addThisTaskButton), withText("Submit"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withId(android.R.id.content),
-                                        0),
-                                5),
-                        isDisplayed()));
         materialButton2.perform(click());
 
+        pressBack();
+
         ViewInteraction textView = onView(
-                allOf(withId(R.id.fragTaskListFragment), withText("1:\n Feed Dawg [NEW]: \nHe is starving!"),
-                        withParent(withParent(withId(R.id.homeTaskListRecyclerView))),
+                allOf(withId(R.id.homeTasksListTitleTextVIew), withText("Jon"),
+                        withParent(withParent(withId(android.R.id.content))),
                         isDisplayed()));
-        textView.check(matches(withText("1:  Feed Dawg [NEW]:  He is starving!")));
+        textView.check(matches(withText("Jon")));
     }
 
     private static Matcher<View> childAtPosition(
