+++ app/src/main/AndroidManifest.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.example.taskmaster">
+    package="com.jrmz.taskmaster">
 
     <application
-        android:name="TaskMasterApplication"
+        android:name="com.jrmz.taskmaster.TaskMasterApplication"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -15,16 +15,16 @@
         android:theme="@style/Theme.Taskmaster"
         tools:targetApi="31">
         <activity
-            android:name=".activites.TaskDetailActivity"
+            android:name="com.jrmz.taskmaster.activites.TaskDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".fragments.UserSettings"
+            android:name="com.jrmz.taskmaster.fragments.UserSettings"
             android:exported="false" />
         <activity
-            android:name=".activites.AddTask"
+            android:name="com.jrmz.taskmaster.activites.AddTask"
             android:exported="false" />
         <activity
-            android:name=".HomeActivity"
+            android:name="com.jrmz.taskmaster.HomeActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
