+++ app/src/main/java/com/jrmz/taskmaster/activites/AddTask.java
@@ -1,4 +1,4 @@
-package com.example.taskmaster.activites;
+package com.jrmz.taskmaster.activites;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -16,7 +16,7 @@ import com.amplifyframework.api.graphql.model.ModelQuery;
 import com.amplifyframework.core.Amplify;
 import com.amplifyframework.datastore.generated.model.Task;
 import com.amplifyframework.datastore.generated.model.Team;
-import com.example.taskmaster.R;
+import com.jrmz.taskmaster.R;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -35,6 +35,7 @@ public class AddTask extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
+        //setContentView(R.layout.activity_add_task);
         setContentView(R.layout.activity_add_task);
         teamsFuture = new CompletableFuture<>();
         this.getTeamsFromDB();
