+++ app/src/main/java/com/jrmz/taskmaster/models/TaskModel.java
@@ -1,4 +1,4 @@
-package com.example.taskmaster.models;
+package com.jrmz.taskmaster.models;
 
 import androidx.annotation.NonNull;
