+++ app/src/main/res/layout/activity_add_task.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/Azure"
-    tools:context=".activites.AddTask">
+    tools:context="com.jrmz.taskmaster.activites.AddTask">
 
     <TextView
         android:id="@+id/addTaskTitleTextView"
