+++ app/src/main/res/layout/activity_home.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/Azure"
-    tools:context=".HomeActivity">
+    tools:context="com.jrmz.taskmaster.HomeActivity">
 
     <TextView
         android:id="@+id/homeTasksListTitleTextVIew"
