+++ app/src/main/res/layout/activity_user_settings.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/Azure"
-    tools:context=".fragments.UserSettings">
+    tools:context="com.jrmz.taskmaster.fragments.UserSettings">
 
     <TextView
         android:id="@+id/enterYourUsernameTextview"
