+++ app/src/main/res/layout/fragment_task_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
-    tools:context=".fragments.TaskListFragment">
+    tools:context="com.jrmz.taskmaster.fragments.TaskListFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
